scripts/requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# AI Summary
google-generativeai>=0.8.0

# X (Twitter) - Optional
twikit>=2.0.0
aiofiles>=23.0.0
